+ integer version
git-svn-id: https://svn.r-project.org/R/trunk@18843 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/qsort-body.c
@@ -17,7 +17,7 @@
      * originally k = 20 -> n_max =    2'097'151
      * now        k = 31 -> n_max = 4294'967'295
      */
-    double vt, vtt;
+    NUMERIC vt, vtt;
     int ii, ij, k, l, m;
 #ifdef qsort_Index
     int it, tt;
